+++ src/trsp/trsp.c
@@ -161,7 +161,7 @@ _pgr_trspv4(PG_FUNCTION_ARGS) {
         int64_t path_seq = call_cntr == 0?  1 : result_tuples[call_cntr - 1].start_id;
 
         values[0] = Int32GetDatum((int32_t)call_cntr + 1);
-        values[1] = Int32GetDatum(path_seq);
+        values[1] = Int32GetDatum((int32_t)path_seq);
         values[2] = Int64GetDatum(result_tuples[call_cntr].start_id);
         values[3] = Int64GetDatum(result_tuples[call_cntr].end_id);
         values[4] = Int64GetDatum(result_tuples[call_cntr].node);
@@ -257,7 +257,7 @@ _v4trsp(PG_FUNCTION_ARGS) {
         int64_t path_seq = call_cntr == 0?  1 : result_tuples[call_cntr - 1].start_id;
 
         values[0] = Int32GetDatum((int32_t)call_cntr + 1);
-        values[1] = Int32GetDatum(path_seq);
+        values[1] = Int32GetDatum((int32_t)path_seq);
         values[2] = Int64GetDatum(result_tuples[call_cntr].start_id);
         values[3] = Int64GetDatum(result_tuples[call_cntr].end_id);
         values[4] = Int64GetDatum(result_tuples[call_cntr].node);
@@ -341,8 +341,8 @@ _trsp(PG_FUNCTION_ARGS) {
 
         int64_t path_seq = call_cntr == 0?  1 : result_tuples[call_cntr - 1].start_id;
 
-        values[0] = Int32GetDatum(call_cntr + 1);
-        values[1] = Int32GetDatum(path_seq);
+        values[0] = Int32GetDatum((int32_t)call_cntr + 1);
+        values[1] = Int32GetDatum((int32_t)path_seq);
         values[2] = Int64GetDatum(result_tuples[call_cntr].start_id);
         values[3] = Int64GetDatum(result_tuples[call_cntr].end_id);
         values[4] = Int64GetDatum(result_tuples[call_cntr].node);
